refactor(wave7b-step2): mechanically split loader behind stable facade
