+++ crates/assay-evidence/src/lint/packs/loader.rs
@@ -7,7 +7,6 @@
 //! - Computes deterministic digest: sha256(JCS(JSON(yaml)))
 
 use super::schema::{PackDefinition, PackValidationError};
-use sha2::{Digest, Sha256};
 use std::path::{Path, PathBuf};
 use thiserror::Error;
 
@@ -93,367 +92,39 @@ pub enum PackError {
 
 /// Load a pack from a reference (file path or built-in name).
 pub fn load_pack(reference: &str) -> Result<LoadedPack, PackError> {
-    let path = Path::new(reference);
-
-    // 1. If path exists on filesystem → load as file or dir
-    if path.exists() {
-        if path.is_dir() {
-            let pack_yaml = path.join("pack.yaml");
-            if pack_yaml.exists() {
-                return load_pack_from_file(&pack_yaml);
-            }
-            // If dir exists but no pack.yaml, we fall through?
-            // User requested: "for directories: pack.yaml required".
-            // If the user explicitly passed a path, we should probably error if it's a dir without pack.yaml,
-            // UNLESS it also happens to match a built-in name?
-            // "Path step" accepts any existing path.
-            // Let's strict fail if it's a directory and explicit path was likely intended?
-            // But wait, if I run `assay lint --pack ./my-pack`, and `./my-pack` is a dir, I expect it to load `pack.yaml`.
-            // If `pack.yaml` is missing, should I fail or check built-ins?
-            // The requirement: "Precedence: Direct File Path > Built-in > Local".
-            // If `reference` resolves to a filesystem path, it wins.
-            // If it's a directory without `pack.yaml`, it's an invalid pack path.
-            return Err(PackError::ReadError {
-                path: pack_yaml,
-                source: std::io::Error::new(
-                    std::io::ErrorKind::NotFound,
-                    "Directory provided but 'pack.yaml' not found",
-                ),
-            });
-        }
-        return load_pack_from_file(path);
-    }
-
-    // 2. Check built-in packs by name
-    if let Some((builtin_name, content)) = get_builtin_pack_with_name(reference) {
-        return load_pack_from_string(content, PackSource::BuiltIn(builtin_name));
-    }
-
-    // 3. Local pack directory (valid name only; containment enforced in load)
-    if is_valid_pack_name(reference) {
-        if let Some(loaded) = try_load_from_config_dir(reference)? {
-            return Ok(loaded);
-        }
-    }
-
-    // 4. Registry / BYOS (future)
-
-    // 5. Not found
-    Err(PackError::NotFound {
-        reference: reference.to_string(),
-        suggestion: suggest_similar_pack(reference),
-    })
-}
-
-/// Look up a built-in pack by name, returning both name and content.
-fn get_builtin_pack_with_name(name: &str) -> Option<(&'static str, &'static str)> {
-    super::BUILTIN_PACKS
-        .iter()
-        .find(|(n, _)| *n == name)
-        .map(|(n, c)| (*n, *c))
-}
-
-/// Try to load a pack from the local config directory.
-fn try_load_from_config_dir(name: &str) -> Result<Option<LoadedPack>, PackError> {
-    let config_dir = match get_config_pack_dir() {
-        Some(d) => d,
-        None => return Ok(None),
-    };
-
-    if !config_dir.exists() {
-        return Ok(None);
-    }
-
-    // Harden canonicalization: if dir exists, must canonicalize or error
-    let canonical_config =
-        std::fs::canonicalize(&config_dir).map_err(|e| PackError::ReadError {
-            path: config_dir.clone(),
-            source: e,
-        })?;
-
-    // Candidates:
-    // 1. {config_dir}/{name}.yaml
-    // 2. {config_dir}/{name}/pack.yaml
-    let candidates = vec![
-        config_dir.join(format!("{}.yaml", name)),
-        config_dir.join(name).join("pack.yaml"),
-    ];
-
-    for path in candidates {
-        if path.exists() {
-            // SECURITY: Path containment check
-            // Use canonicalize() to resolve symlinks and '..'
-            let canonical_path =
-                std::fs::canonicalize(&path).map_err(|e| PackError::ReadError {
-                    path: path.clone(),
-                    source: e,
-                })?;
-
-            if !canonical_path.starts_with(&canonical_config) {
-                // Determine if this is a symlink escape attempt
-                return Err(PackValidationError::Safety(format!(
-                    "Pack path '{}' resolves outside config directory '{}'",
-                    path.display(),
-                    config_dir.display()
-                ))
-                .into());
-            }
-
-            // Mitigate TOCTOU: load from the canonical path we just verified
-            return Ok(Some(load_pack_from_file(&canonical_path)?));
-        }
-    }
-
-    Ok(None)
-}
-
-/// Determine the config pack directory per ADR-021.
-fn get_config_pack_dir() -> Option<PathBuf> {
-    // Unix: $XDG_CONFIG_HOME/assay/packs or ~/.config/assay/packs
-    // Windows: %APPDATA%\assay\packs
-
-    #[cfg(not(windows))]
-    {
-        if let Ok(xdg) = std::env::var("XDG_CONFIG_HOME") {
-            if !xdg.is_empty() {
-                return Some(PathBuf::from(xdg).join("assay").join("packs"));
-            }
-        }
-
-        // Fallback: ~/.config/assay/packs
-        if let Ok(home) = std::env::var("HOME") {
-            return Some(
-                PathBuf::from(home)
-                    .join(".config")
-                    .join("assay")
-                    .join("packs"),
-            );
-        }
-    }
-
-    #[cfg(windows)]
-    {
-        if let Ok(appdata) = std::env::var("APPDATA") {
-            return Some(PathBuf::from(appdata).join("assay").join("packs"));
-        }
-    }
-
-    None
+    loader_internal::run::load_pack_impl(reference)
 }
 
 /// Validate pack name grammar per ADR-021.
 ///
 /// Chars: a-z, 0-9, -
 /// Must not start or end with hyphen.
+#[cfg(test)]
 fn is_valid_pack_name(name: &str) -> bool {
-    if name.is_empty() {
-        return false;
-    }
-    if name.starts_with('-') || name.ends_with('-') {
-        return false;
-    }
-    name.chars()
-        .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '-')
+    loader_internal::resolve::is_valid_pack_name_impl(name)
 }
 
 /// Load multiple packs from references.
 pub fn load_packs(references: &[String]) -> Result<Vec<LoadedPack>, PackError> {
-    let mut packs = Vec::with_capacity(references.len());
-    for reference in references {
-        packs.push(load_pack(reference)?);
-    }
-    Ok(packs)
+    loader_internal::run::load_packs_impl(references)
 }
 
 /// Load a pack from a file path.
 pub fn load_pack_from_file(path: &Path) -> Result<LoadedPack, PackError> {
-    let content = std::fs::read_to_string(path).map_err(|e| PackError::ReadError {
-        path: path.to_path_buf(),
-        source: e,
-    })?;
-
-    load_pack_from_string(&content, PackSource::File(path.to_path_buf()))
-}
-
-/// Load pack from string content.
-///
-/// Implements strict YAML parsing for the Pack Engine v1 spec:
-/// - Rejects duplicate keys (partially via serde map handling)
-/// - Rejects unknown fields (`deny_unknown_fields`)
-///
-/// Note: While the spec discourages anchors/aliases, the current implementation
-/// accepts them if they resolve to valid JSON. Future versions may enforce failing on anchors/aliases.
-fn load_pack_from_string(content: &str, source: PackSource) -> Result<LoadedPack, PackError> {
-    // Parse YAML with strict settings (deny_unknown_fields on schema types)
-    let definition: PackDefinition =
-        serde_yaml::from_str(content).map_err(|e| PackError::YamlParseError {
-            message: format_yaml_error(e),
-        })?;
-
-    // Validate the pack
-    definition.validate()?;
-
-    // Check version compatibility
-    check_version_compatibility(&definition)?;
-
-    // Compute digest
-    let digest = compute_pack_digest(&definition)?;
-
-    Ok(LoadedPack {
-        definition,
-        digest,
-        source,
-    })
-}
-
-/// Compute pack digest: sha256(JCS(JSON(pack)))
-fn compute_pack_digest(definition: &PackDefinition) -> Result<String, PackError> {
-    // Serialize to canonical JSON using RFC 8785 JCS
-    let canonical = serde_jcs::to_string(definition).map_err(|e| PackError::YamlParseError {
-        message: format!("Failed to canonicalize pack to JCS JSON: {}", e),
-    })?;
-
-    // Compute SHA-256
-    let hash = Sha256::digest(canonical.as_bytes());
-    Ok(format!("sha256:{}", hex::encode(hash)))
-}
-
-/// Check if current Assay version satisfies pack requirements.
-fn check_version_compatibility(definition: &PackDefinition) -> Result<(), PackError> {
-    let current_version = env!("CARGO_PKG_VERSION");
-    let required = &definition.requires.assay_min_version;
-
-    // Parse the required version constraint
-    // Format: ">=X.Y.Z" or just "X.Y.Z"
-    let required_version = required.trim_start_matches(">=").trim_start_matches("=");
-
-    // Simple semver comparison (major.minor.patch)
-    if !version_satisfies(current_version, required_version) {
-        return Err(PackError::IncompatibleVersion {
-            pack: definition.name.clone(),
-            required: required.clone(),
-            current: current_version.to_string(),
-        });
-    }
-
-    Ok(())
+    loader_internal::run::load_pack_from_file_impl(path)
 }
 
 /// Simple semver comparison (checks if current >= required).
+#[cfg(test)]
 fn version_satisfies(current: &str, required: &str) -> bool {
-    let parse_version = |v: &str| -> Option<(u32, u32, u32)> {
-        let parts: Vec<&str> = v.split('.').collect();
-        if parts.len() >= 3 {
-            Some((
-                parts[0].parse().ok()?,
-                parts[1].parse().ok()?,
-                parts[2].split('-').next()?.parse().ok()?,
-            ))
-        } else if parts.len() == 2 {
-            Some((parts[0].parse().ok()?, parts[1].parse().ok()?, 0))
-        } else {
-            None
-        }
-    };
-
-    match (parse_version(current), parse_version(required)) {
-        (Some((c_major, c_minor, c_patch)), Some((r_major, r_minor, r_patch))) => {
-            (c_major, c_minor, c_patch) >= (r_major, r_minor, r_patch)
-        }
-        _ => false, // If we can't parse, treat as incompatible (fail closed)
-    }
-}
-
-/// Format YAML parsing error for user-friendly display.
-fn format_yaml_error(e: serde_yaml::Error) -> String {
-    let msg = e.to_string();
-
-    // Check for duplicate key errors
-    if msg.contains("duplicate key") {
-        return format!(
-            "Duplicate key detected (duplicate keys not allowed for security): {}",
-            msg
-        );
-    }
-
-    // Check for unknown field errors
-    if msg.contains("unknown field") {
-        return format!(
-            "Unknown field detected (prevents digest bypass attacks): {}",
-            msg
-        );
-    }
-
-    msg
-}
-
-/// Suggest similar pack names.
-fn suggest_similar_pack(reference: &str) -> String {
-    use super::BUILTIN_PACKS;
-
-    // Simple prefix matching for suggestions
-    let suggestions: Vec<&str> = BUILTIN_PACKS
-        .iter()
-        .filter(|(name, _)| {
-            name.starts_with(reference)
-                || reference.starts_with(*name)
-                || levenshtein_distance(name, reference) <= 3
-        })
-        .map(|(name, _)| *name)
-        .collect();
-
-    if suggestions.is_empty() {
-        format!(
-            "Available built-in packs: {}. Or specify a file path: --pack ./my-pack.yaml",
-            BUILTIN_PACKS
-                .iter()
-                .map(|(n, _)| *n)
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    } else {
-        format!("Did you mean '{}'?", suggestions.join("' or '"))
-    }
+    loader_internal::compat::version_satisfies_impl(current, required)
 }
 
 /// Simple Levenshtein distance for fuzzy matching.
 #[allow(clippy::needless_range_loop)]
+#[cfg(test)]
 fn levenshtein_distance(a: &str, b: &str) -> usize {
-    let a_chars: Vec<char> = a.chars().collect();
-    let b_chars: Vec<char> = b.chars().collect();
-    let m = a_chars.len();
-    let n = b_chars.len();
-
-    if m == 0 {
-        return n;
-    }
-    if n == 0 {
-        return m;
-    }
-
-    let mut dp = vec![vec![0usize; n + 1]; m + 1];
-
-    for i in 0..=m {
-        dp[i][0] = i;
-    }
-    for j in 0..=n {
-        dp[0][j] = j;
-    }
-
-    for i in 1..=m {
-        for j in 1..=n {
-            let cost = if a_chars[i - 1] == b_chars[j - 1] {
-                0
-            } else {
-                1
-            };
-            dp[i][j] = (dp[i - 1][j] + 1)
-                .min(dp[i][j - 1] + 1)
-                .min(dp[i - 1][j - 1] + cost);
-        }
-    }
-
-    dp[m][n]
+    loader_internal::resolve::levenshtein_distance_impl(a, b)
 }
 
 #[cfg(test)]
